Make shared CLI and pull some stuff out into Core
