+++ src/Nitro/CommandLine/src/CommandLine.Cloud/Commands/FusionConfiguration/PublishCommand/FusionConfigurationPublishStartCommand.cs
@@ -1,6 +1,5 @@
 using System.CommandLine.Invocation;
 using ChilliCream.Nitro.CLI.Client;
-using ChilliCream.Nitro.CLI.Exceptions;
 using ChilliCream.Nitro.CLI.Option;
 using ChilliCream.Nitro.CLI.Option.Binders;
