+++ src/Nitro/CommandLine/src/CommandLine.Cloud/Commands/Schema/UploadSchemaCommand.cs
@@ -1,5 +1,4 @@
 using ChilliCream.Nitro.CLI.Client;
-using ChilliCream.Nitro.CLI.Exceptions;
 using ChilliCream.Nitro.CLI.Helpers;
 using ChilliCream.Nitro.CLI.Option;
 using ChilliCream.Nitro.CLI.Option.Binders;
