+++ src/Nitro/CommandLine/src/CommandLine.Cloud/Commands/Workspace/CreateWorkspaceCommand.cs
@@ -1,7 +1,6 @@
 using System.CommandLine.Invocation;
 using ChilliCream.Nitro.CLI.Auth;
 using ChilliCream.Nitro.CLI.Client;
-using ChilliCream.Nitro.CLI.Exceptions;
 using ChilliCream.Nitro.CLI.Option;
 using ChilliCream.Nitro.CLI.Option.Binders;
 using ChilliCream.Nitro.CLI.Results;
