+++ src/Nitro/CommandLine/src/CommandLine.Cloud/Commands/Workspace/SetDefaultWorkspaceCommand.cs
@@ -1,6 +1,5 @@
 using ChilliCream.Nitro.CLI.Auth;
 using ChilliCream.Nitro.CLI.Client;
-using ChilliCream.Nitro.CLI.Exceptions;
 using ChilliCream.Nitro.CLI.Option.Binders;
 using static ChilliCream.Nitro.CLI.ThrowHelper;
