+++ src/Nitro/CommandLine/src/CommandLine.Cloud/Extensions/ApiClientCommandLineBuilderExtensions.cs
@@ -2,7 +2,6 @@ using System.CommandLine.Builder;
 using System.CommandLine.Parsing;
 using System.Net.Http.Headers;
 using ChilliCream.Nitro.CLI.Client;
-using ChilliCream.Nitro.CLI.Exceptions;
 using Microsoft.Net.Http.Headers;
 using static System.UriKind;
 
@@ -11,7 +10,7 @@ namespace ChilliCream.Nitro.CLI.Option.Binders;
 internal static class ApiClientCommandLineBuilderExtensions
 {
     private static readonly string _userAgent = $"Nitro CLI/{Version}";
-    private const string _clientId = "<<NITRO_GRAPHQL_CLIENT_ID>>"; // TODO inject client id via build
+    private const string _clientId = "<<NITRO_GRAPHQL_CLIENT_ID>>"; // TODO: inject client id via build
 
     public static CommandLineBuilder AddApiClient(this CommandLineBuilder builder)
         => builder
