+++ src/Nitro/CommandLine/src/CommandLine.Cloud/Helpers/ConsoleHelpers.cs
@@ -1,143 +1,10 @@
-using System.CommandLine.Invocation;
-using System.Text.Json;
 using ChilliCream.Nitro.CLI.Client;
-using ChilliCream.Nitro.CLI.Exceptions;
-using ChilliCream.Nitro.CLI.Results;
-using Spectre.Console.Json;
 using StrawberryShake;
 
 namespace ChilliCream.Nitro.CLI;
 
 internal static class ConsoleHelpers
 {
-    public static void PrintHeader(this IAnsiConsole console)
-    {
-        console.Write(new FigletText("Nitro").Color(Color.Red).Centered());
-        console.Write(new Rule("[white]By ChilliCream[/]"));
-        console.WriteLine();
-    }
-
-    public static void Log(this IAnsiConsole console, string str)
-    {
-        console.MarkupLine("[grey]LOG: [/]" + str);
-    }
-
-    public static Status DefaultStatus(this IAnsiConsole console)
-    {
-        return console.Status()
-            .Spinner(Spinner.Known.BouncingBar)
-            .SpinnerStyle(Style.Parse("green bold"));
-    }
-
-    public static void Title(this IAnsiConsole console, string str)
-    {
-        console.MarkupLineInterpolated($"[white bold]{str}:[/]");
-        console.WriteLine();
-    }
-
-    public static void Success(this IAnsiConsole console, string message)
-    {
-        console.MarkupLine($"[green bold]{message}[/]");
-    }
-
-    public static void OkLine(this IAnsiConsole console, string message)
-    {
-        console.MarkupLine(Glyphs.Check.Space() + message);
-    }
-
-    public static void Json(this IAnsiConsole console, object obj)
-    {
-        // TODO enable for full AOT compliancy
-        // console.Write(new JsonText(
-        //     JsonSerializer.Serialize(obj,
-        //         new JsonSerializerOptions { PropertyNamingPolicy = JsonNamingPolicy.CamelCase })));
-        console.WriteLine("NOT SUPPORTED");
-    }
-
-    public static void ErrorLine(this IAnsiConsole console, string message)
-    {
-        console.MarkupLine(Glyphs.Cross.Space() + message);
-    }
-
-    public static void OkQuestion(this IAnsiConsole console, string question, string result)
-    {
-        console.MarkupLine(
-            $"{Glyphs.QuestionMark.Space()}[bold]{question}[/]: [darkseagreen4]{result}[/]");
-    }
-
-    public static async Task<string> AskAsync(
-        this IAnsiConsole console,
-        string question,
-        CancellationToken cancellationToken)
-        => await new TextPrompt<string>(question.AsQuestion())
-            .ShowAsync(console, cancellationToken);
-
-    public static async Task<string> OptionOrAskAsync(
-        this InvocationContext context,
-        string question,
-        Option<string> option,
-        CancellationToken cancellationToken)
-    {
-        var value = context.ParseResult.GetValueForOption(option);
-
-        if (value is not null)
-        {
-            return value;
-        }
-
-        var console = context.BindingContext.GetRequiredService<IAnsiConsole>();
-
-        return await new TextPrompt<string>(question.AsQuestion())
-            .ShowAsync(console, cancellationToken);
-    }
-
-    public static async Task<string> AskAsync(
-        this IAnsiConsole console,
-        string question,
-        string defaultValue,
-        CancellationToken cancellationToken)
-    {
-        var questionText = $"{question}".AsQuestion();
-        var prompt = new TextPrompt<string>(questionText).DefaultValue(defaultValue);
-        return await prompt.ShowAsync(console, cancellationToken);
-    }
-
-    public static async Task<bool> ConfirmAsync(
-        this IAnsiConsole console,
-        string question,
-        CancellationToken cancellationToken)
-        => await new ConfirmationPrompt(question.AsQuestion())
-            .ShowAsync(console, cancellationToken);
-
-    public static async Task<bool> OptionOrConfirmAsync(
-        this InvocationContext context,
-        string question,
-        Option<bool?> option,
-        CancellationToken cancellationToken)
-    {
-        var value = context.ParseResult.GetValueForOption(option);
-
-        if (value is not null)
-        {
-            return value.Value;
-        }
-
-        var console = context.BindingContext.GetRequiredService<IAnsiConsole>();
-
-        return await new ConfirmationPrompt(question.AsQuestion())
-            .ShowAsync(console, cancellationToken);
-    }
-
-    public static void WarningLine(this IAnsiConsole console, string message)
-    {
-        console.MarkupLine(Glyphs.ExclamationMark.Space() + message);
-    }
-
-    public static void Error(this IAnsiConsole console, string message)
-    {
-        console.MarkupLine($"[red bold]{message}[/]");
-    }
-
     public static void EnsureNoErrors<T>(
         this IAnsiConsole console,
         IOperationResult<T> result) where T : class
@@ -160,29 +27,6 @@ internal static class ConsoleHelpers
         }
 
         return result.Data;
-    }
-
-    public static void PrintError(this IAnsiConsole console, string message, string? code = null)
-    {
-        if (code is not null)
-        {
-            console.MarkupLineInterpolated(
-                $"[red][bold]Error[/]: {message}[/][grey] ({code})[/]");
-        }
-        else
-        {
-            console.MarkupLineInterpolated($"[red][bold]Error[/]: {message}[/]");
-        }
-    }
-
-    public static bool IsHumandReadable(this IAnsiConsole console)
-    {
-        return console is IExtendedConsole { IsInteractive: true };
-    }
-
-    public static IDisposable UseInteractive(this IAnsiConsole console)
-    {
-        return new InteractiveScope(console);
     }
 
     public static void PrintErrorsAndExit<T>(this IAnsiConsole console, IReadOnlyList<T>? errors)
@@ -418,31 +262,6 @@ internal static class ConsoleHelpers
             default:
                 ansiConsole.Error("Unexpected Error");
                 break;
-        }
-    }
-
-    private sealed class InteractiveScope : IDisposable
-    {
-        private readonly IAnsiConsole _console;
-        private readonly bool _originalValue;
-
-        public InteractiveScope(IAnsiConsole console)
-        {
-            _console = console;
-
-            if (_console is IExtendedConsole customConsole)
-            {
-                _originalValue = customConsole.IsInteractive;
-                customConsole.IsInteractive = true;
-            }
-        }
-
-        public void Dispose()
-        {
-            if (_console is IExtendedConsole customConsole)
-            {
-                customConsole.IsInteractive = _originalValue;
-            }
         }
     }
 }
