+++ src/Nitro/CommandLine/src/CommandLine.Cloud/Nitro.CommandLine.Cloud.csproj
@@ -140,6 +140,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\..\..\StrawberryShake\MetaPackages\Server\StrawberryShake.Server.csproj" />
+    <ProjectReference Include="..\CommandLine.Core\Nitro.CommandLine.Core.csproj" />
     <ProjectReference Include="..\CommandLine.Fusion.Compatibility\Nitro.CommandLine.Fusion.Compatibility.csproj" />
   </ItemGroup>
 </Project>
