--- src/Nitro/CommandLine/src/CommandLine.Cloud/Program.cs
@@ -1,4 +1,0 @@
-using System.CommandLine.Parsing;
-using ChilliCream.Nitro.CLI;
-
-// return await new NitroCommandLine().Build().InvokeAsync(args);
