--- src/Nitro/CommandLine/src/CommandLine.Cloud/Result/IExtendedConsole.cs
@@ -1,6 +1,0 @@
-namespace ChilliCream.Nitro.CLI.Results;
-
-internal interface IExtendedConsole : IAnsiConsole
-{
-    public bool IsInteractive { get; set; }
-}
