+++ src/Nitro/CommandLine/src/CommandLine.Core/Helpers/ExitCodes.cs
@@ -1,0 +1,8 @@
+namespace ChilliCream.Nitro;
+
+public static class ExitCodes
+{
+    public const int Success = 0;
+
+    public const int Error = -1;
+}
