+++ src/Nitro/CommandLine/src/CommandLine.Core/Helpers/Glyphs.cs
@@ -1,0 +1,12 @@
+namespace ChilliCream.Nitro;
+
+internal static class Glyphs
+{
+    public const string Check = "[green bold]✓[/]";
+
+    public const string Cross = "[red bold]✕[/]";
+
+    public const string QuestionMark = "[lime bold]?[/]";
+
+    public const string ExclamationMark = "[yellow bold]![/]";
+}
