+++ src/Nitro/CommandLine/src/CommandLine.Core/Nitro.CommandLine.Core.csproj
@@ -5,4 +5,10 @@
     <AssemblyName>Nitro.CommandLine.Core</AssemblyName>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" />
+    <PackageReference Include="Spectre.Console" />
+    <PackageReference Include="System.CommandLine" />
+  </ItemGroup>
+
 </Project>
