+++ src/Nitro/CommandLine/src/CommandLine.Core/Result/IExtendedConsole.cs
@@ -1,0 +1,6 @@
+namespace ChilliCream.Nitro;
+
+public interface IExtendedConsole : IAnsiConsole
+{
+    public bool IsInteractive { get; set; }
+}
