+++ src/Nitro/CommandLine/src/CommandLine.Core/Usings.cs
@@ -1,0 +1,3 @@
+global using System.CommandLine;
+global using Microsoft.Extensions.DependencyInjection;
+global using Spectre.Console;
