+++ src/Nitro/CommandLine/src/CommandLine/NitroRootCommand.cs
@@ -1,0 +1,5 @@
+using System.CommandLine;
+
+namespace ChilliCream.Nitro.CommandLine;
+
+public sealed class NitroRootCommand() : Command("nitro");
